api security enhanced

+++ app/api/v1/purchase/route.js
@@ -2,8 +2,17 @@ import { connectToDB } from "@/app/lib/dbconnect";
 import { Products } from "@/app/lib/Models/productSchema";
 import { NextResponse } from "next/server";
 
-export async function GET(req) {
+const Token = process.env.TOKEN;
+
+export async function DELETE(req) {
+  const { token } = await req.json();
   try {
+    if (token != Token) {
+      return NextResponse.json({
+        message: "Unauthorized Token",
+        success: false,
+      });
+    }
     await connectToDB();
     const products = await Products.deleteMany();
     return NextResponse.json({
